Update CMakeLists.txt
Update FFmpeg from 1.2.2 to FFmpeg 1.2.4 "Magic"

1.2.4 was released on 2013-10-03. It is the latest stable FFmpeg release from the 1.2 release branch, which was cut from master on 2013-03-07. Amongst lots of other changes, it includes all changes from ffmpeg-mt, libav master of 2013-03-07, libav 9 as of 2013-03-07. 
It includes the following library versions:
  libavutil      52. 18.100
  libavcodec     54. 92.100
  libavformat    54. 63.104
  libavdevice    54.  3.103
  libavfilter     3. 42.103
  libswscale      2.  2.100
  libswresample   0. 17.102
  libpostproc    52.  2.100

Magic, well funnily enough it does seem to be "magic" as the blank video problem seems to have disappeared? :D

I will look into incorporating the newer FFmpeg version 2.0, not as easy to do as it requires a lot of patching. 1.2.4 is a step in the right direction.