Update Repository.cpp
This would explain the error I got when pushing the merge. It said there was an error with this file and it seems a part of the code change didn't make it. I have found the problem and fixed it and the compilation completed. I will test the build to confirm everything is in order.