Avoid dangling unsubmitted frames from previous video for #62.
Kudos to @quasar1, your post in
http://boxeed.in/forums/viewtopic.php?f=14&t=362 reminded me that
I've seen it quite often in the case where there is a short commercial
video followed by the "real" video, and it does indeed reproduce every
single time for me.

With a proper test case I was able to track it down. Hopefully this
fixes all of the "blank screen at start" issues, but more testing is
needed.