Add libxslt and libxml2
Only problem with this was finding the libcrypt-config, hence the 
ac_cv_path_LIBGCRYPT_CONFIG=/usr/local/bin/
Hopefully everyone else can build with this, otherwise we may have to tweak it a bit.
Had to update libcrypt as research told me libxslt needed libcrypt to be version 1.6.0
I will make a new thread on the forums so we can iron out any issues if there are any. Currently I have just built without any problems, need to give it a test now.