Revert lib names except libenca and libass
Otherwise XBMC gets her panties in a bunch! (Launching of videos stall and cause GFX blocking).
Unsure as to why this is but it's evident that searching the xbmc.bin for the correct lib name is not the answer!
This should help you guys out there. I would never walk away leaving a problem I 'contributed' to and let you pull your hair out trying to fix it. Hence why when I discovered the problem I spent two days tracking the changes.