Revert libenca and libass changes
Well, I tried everything but I can't get libenca to work anymore. It must have been a fluke it worked when the lib name was changed. Now when searching the bin it isn't found at all.
Not even sure XBMC is meant to look for it has it's not mentioned in the configure.in or even the config.log.
Since I don't want it or need it I'm throwing in the towel! :)