Do not export sqlite lib
Seems the sqlite module that python builds interferes with the library file XBMC links to.
Python 2.6 and up have sqlite built in via modules, so exporting the library file isn't needed.
The stalling and gfx hitching when launching videos through the menu was because two instances of sqlite were being initiated. Well that's what looks to be happening from where I'm standing.
The correct lib naming triggered this, then changing the lib name back cured it. So the best course of action is not not export the lib at all.