Add SSL support to curl
Required making Openssl a shared library, previously we never had ssl
support in curl because the ssl library was compiled statically and was
never found.
Rtmpdump needs ssl enabling, but not really that important as it never
had ssl support before. Disabled for now to enable rtmpdump to compile
to completion.