StdString.h - comment out offending lines which cause compilation errors.
This is the best I can come up with for now.
Thanks to KeyserSoze1 for his assistance in tracking the issue!
Tested the build for a couple of weeks and I didn't see any issues due
to these changes. If anyone can do better get your fingers out of your
ass and give us a hand (after you washed it) :D