libcec: remove from project
Kodi requires a newer libcec version which cannot be compiled with our
current toolchain. The toolchain can be upgraded and I have previously
done this but the catch is the compiled bin and libs will not run on the
Boxee's antiquated kernel/glibc (not compatible).
I don't think anyone got libcec working on their Boxee anyway, so time
we ditched it...