Experimental elastic video codec buffer batching.
This allows queueing up buffers to ~75% of total queue possible, and
should be much better at feeding the hardware pipe with data at a
rate close to what it is capable of processing.

In other words: if it detects it's feeding the hw too slow it'll try
to feed it more data and if it detects it's feeding it too fast it'll
start feeding it data at a slower rate.