FFmpeg: downgrade to 2.4.4-Helix & re--enable-asm
re--enable-asm (we need this enabled as video output is sluggish with it
disabled).

downgrade to 2.4.4-Helix (fixes broken thumbnail extraction on h.264
video files).
I thought it was the compiler at first but further investigation proved
otherwise.

the borked subtitles seem no worse, maybe slightly better. I'm waiting
for a Kodi PR on dirty regions which may address the issue.